Move ClassDefinitionPrinters loading later in the bootstrap process
ClassDefinitionPrinters package is currently loaded by Hermes alongside the compiler. But this is needed only for tooling, the compiler does not need it except in the UI  package. 

I propose to load it later in the bootstrap process so that it gets loaded by Metacello normaly and not by Hermes