Calypso: Allow definition of shared variables when using the class definition tab as a regular code editor
For this change we make ClyTextEditingMode ask the browser tool if it is in "scripting" or "method" mode.
We also remove the cached `isForScripting` instance variable.

FInally we extract the current heuristic for wether the code of a class definition tab refers to a class defition or a method into #codeLooksLikeAClassDefinition:

Co-Authored-By: Matija Kljajic <matijakljajic173@gmail.com>