Only pass background exceptions when we #shouldPassBackgroundFailures (fixes #17813)
Normally during test cleanup we will terminate any suspended background processes, so they will not actually get to resume and hit the #pass. However if the error was raised in an #ensure: block, the background process will finish executing the handler even when sent #terminate, so we need to avoid passing the exception and opening a debugger.